+++ crates/vm/src/arch/execution_mode/mod.rs
@@ -4,5 +4,5 @@ pub mod tracegen;
 
 // TODO(ayush): better name
 pub trait E1E2ExecutionCtx {
-    fn on_memory_operation(&mut self, address_space: u32, ptr: u32, size: usize);
+    fn on_memory_operation(&mut self, address_space: u32, ptr: u32, size: u32);
 }
